requirements.txt
streamlit==1.34.0
pandas==2.2.2
faiss-cpu==1.7.4
sentence-transformers==2.6.1
requests==2.31.0
python-dotenv==1.0.1
